refactor: remove unused generation_logs table and generation_count field
Remove generation_logs table and generation_count field from daily_reports
to simplify the database schema and codebase. These features were defined
but never implemented or used in the application.

Database changes:
- Add Flyway migration V7 to drop generation_logs table
- Remove generation_count column from daily_reports table

Backend changes:
- Remove generationCount field from DailyReport entity and related DTOs
- Update repository layer JOOQ mappings
- Modify use cases to remove generation count handling
- Fix tests to match new structure
- Regenerate JOOQ classes and OpenAPI specification

Frontend changes:
- Remove generation count display from ReportCard component
- Remove generationCount from mock data
- Regenerate TypeScript API types

Documentation:
- Update database design documentation
