+++ backend/src/jooq/com/example/backend/jooq/Tables.java
@@ -6,7 +6,6 @@ package com.example.backend.jooq;
 
 import com.example.backend.jooq.tables.JDailyReports;
 import com.example.backend.jooq.tables.JFlywaySchemaHistory;
-import com.example.backend.jooq.tables.JGenerationLogs;
 import com.example.backend.jooq.tables.JGithubCredentials;
 import com.example.backend.jooq.tables.JNotionCredentials;
 import com.example.backend.jooq.tables.JReportTemplates;
@@ -29,11 +28,6 @@ public class Tables {
      * The table <code>public.flyway_schema_history</code>.
      */
     public static final JFlywaySchemaHistory FLYWAY_SCHEMA_HISTORY = JFlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY;
-
-    /**
-     * The table <code>public.generation_logs</code>.
-     */
-    public static final JGenerationLogs GENERATION_LOGS = JGenerationLogs.GENERATION_LOGS;
 
     /**
      * The table <code>public.github_credentials</code>.
