+++ backend/src/jooq/com/example/backend/jooq/tables/records/JDailyReportsRecord.java
@@ -13,8 +13,8 @@ import java.util.UUID;
 import org.jooq.Field;
 import org.jooq.JSONB;
 import org.jooq.Record1;
-import org.jooq.Record9;
-import org.jooq.Row9;
+import org.jooq.Record8;
+import org.jooq.Row8;
 import org.jooq.impl.UpdatableRecordImpl;
 
 
@@ -22,7 +22,7 @@ import org.jooq.impl.UpdatableRecordImpl;
  * This class is generated by jOOQ.
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord> implements Record9<UUID, UUID, LocalDate, JSONB, String, Integer, String, LocalDateTime, LocalDateTime> {
+public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord> implements Record8<UUID, UUID, LocalDate, JSONB, String, String, LocalDateTime, LocalDateTime> {
 
     private static final long serialVersionUID = 1L;
 
@@ -97,59 +97,45 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     /**
-     * Setter for <code>public.daily_reports.generation_count</code>.
-     */
-    public void setGenerationCount(Integer value) {
-        set(5, value);
-    }
-
-    /**
-     * Getter for <code>public.daily_reports.generation_count</code>.
-     */
-    public Integer getGenerationCount() {
-        return (Integer) get(5);
-    }
-
-    /**
      * Setter for <code>public.daily_reports.additional_notes</code>.
      */
     public void setAdditionalNotes(String value) {
-        set(6, value);
+        set(5, value);
     }
 
     /**
      * Getter for <code>public.daily_reports.additional_notes</code>.
      */
     public String getAdditionalNotes() {
-        return (String) get(6);
+        return (String) get(5);
     }
 
     /**
      * Setter for <code>public.daily_reports.created_at</code>.
      */
     public void setCreatedAt(LocalDateTime value) {
-        set(7, value);
+        set(6, value);
     }
 
     /**
      * Getter for <code>public.daily_reports.created_at</code>.
      */
     public LocalDateTime getCreatedAt() {
-        return (LocalDateTime) get(7);
+        return (LocalDateTime) get(6);
     }
 
     /**
      * Setter for <code>public.daily_reports.updated_at</code>.
      */
     public void setUpdatedAt(LocalDateTime value) {
-        set(8, value);
+        set(7, value);
     }
 
     /**
      * Getter for <code>public.daily_reports.updated_at</code>.
      */
     public LocalDateTime getUpdatedAt() {
-        return (LocalDateTime) get(8);
+        return (LocalDateTime) get(7);
     }
 
     // -------------------------------------------------------------------------
@@ -162,17 +148,17 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     // -------------------------------------------------------------------------
-    // Record9 type implementation
+    // Record8 type implementation
     // -------------------------------------------------------------------------
 
     @Override
-    public Row9<UUID, UUID, LocalDate, JSONB, String, Integer, String, LocalDateTime, LocalDateTime> fieldsRow() {
-        return (Row9) super.fieldsRow();
+    public Row8<UUID, UUID, LocalDate, JSONB, String, String, LocalDateTime, LocalDateTime> fieldsRow() {
+        return (Row8) super.fieldsRow();
     }
 
     @Override
-    public Row9<UUID, UUID, LocalDate, JSONB, String, Integer, String, LocalDateTime, LocalDateTime> valuesRow() {
-        return (Row9) super.valuesRow();
+    public Row8<UUID, UUID, LocalDate, JSONB, String, String, LocalDateTime, LocalDateTime> valuesRow() {
+        return (Row8) super.valuesRow();
     }
 
     @Override
@@ -201,22 +187,17 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     @Override
-    public Field<Integer> field6() {
-        return JDailyReports.DAILY_REPORTS.GENERATION_COUNT;
-    }
-
-    @Override
-    public Field<String> field7() {
+    public Field<String> field6() {
         return JDailyReports.DAILY_REPORTS.ADDITIONAL_NOTES;
     }
 
     @Override
-    public Field<LocalDateTime> field8() {
+    public Field<LocalDateTime> field7() {
         return JDailyReports.DAILY_REPORTS.CREATED_AT;
     }
 
     @Override
-    public Field<LocalDateTime> field9() {
+    public Field<LocalDateTime> field8() {
         return JDailyReports.DAILY_REPORTS.UPDATED_AT;
     }
 
@@ -246,22 +227,17 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     @Override
-    public Integer component6() {
-        return getGenerationCount();
-    }
-
-    @Override
-    public String component7() {
+    public String component6() {
         return getAdditionalNotes();
     }
 
     @Override
-    public LocalDateTime component8() {
+    public LocalDateTime component7() {
         return getCreatedAt();
     }
 
     @Override
-    public LocalDateTime component9() {
+    public LocalDateTime component8() {
         return getUpdatedAt();
     }
 
@@ -291,22 +267,17 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     @Override
-    public Integer value6() {
-        return getGenerationCount();
-    }
-
-    @Override
-    public String value7() {
+    public String value6() {
         return getAdditionalNotes();
     }
 
     @Override
-    public LocalDateTime value8() {
+    public LocalDateTime value7() {
         return getCreatedAt();
     }
 
     @Override
-    public LocalDateTime value9() {
+    public LocalDateTime value8() {
         return getUpdatedAt();
     }
 
@@ -341,31 +312,25 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     }
 
     @Override
-    public JDailyReportsRecord value6(Integer value) {
-        setGenerationCount(value);
-        return this;
-    }
-
-    @Override
-    public JDailyReportsRecord value7(String value) {
+    public JDailyReportsRecord value6(String value) {
         setAdditionalNotes(value);
         return this;
     }
 
     @Override
-    public JDailyReportsRecord value8(LocalDateTime value) {
+    public JDailyReportsRecord value7(LocalDateTime value) {
         setCreatedAt(value);
         return this;
     }
 
     @Override
-    public JDailyReportsRecord value9(LocalDateTime value) {
+    public JDailyReportsRecord value8(LocalDateTime value) {
         setUpdatedAt(value);
         return this;
     }
 
     @Override
-    public JDailyReportsRecord values(UUID value1, UUID value2, LocalDate value3, JSONB value4, String value5, Integer value6, String value7, LocalDateTime value8, LocalDateTime value9) {
+    public JDailyReportsRecord values(UUID value1, UUID value2, LocalDate value3, JSONB value4, String value5, String value6, LocalDateTime value7, LocalDateTime value8) {
         value1(value1);
         value2(value2);
         value3(value3);
@@ -374,7 +339,6 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
         value6(value6);
         value7(value7);
         value8(value8);
-        value9(value9);
         return this;
     }
 
@@ -392,7 +356,7 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
     /**
      * Create a detached, initialised JDailyReportsRecord
      */
-    public JDailyReportsRecord(UUID id, UUID userId, LocalDate reportDate, JSONB rawData, String finalContent, Integer generationCount, String additionalNotes, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public JDailyReportsRecord(UUID id, UUID userId, LocalDate reportDate, JSONB rawData, String finalContent, String additionalNotes, LocalDateTime createdAt, LocalDateTime updatedAt) {
         super(JDailyReports.DAILY_REPORTS);
 
         setId(id);
@@ -400,7 +364,6 @@ public class JDailyReportsRecord extends UpdatableRecordImpl<JDailyReportsRecord
         setReportDate(reportDate);
         setRawData(rawData);
         setFinalContent(finalContent);
-        setGenerationCount(generationCount);
         setAdditionalNotes(additionalNotes);
         setCreatedAt(createdAt);
         setUpdatedAt(updatedAt);
