--- backend/src/jooq/com/example/backend/jooq/tables/records/JGenerationLogsRecord.java
@@ -1,333 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.example.backend.jooq.tables.records;
-
-
-import com.example.backend.jooq.tables.JGenerationLogs;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-import org.jooq.Field;
-import org.jooq.Record1;
-import org.jooq.Record7;
-import org.jooq.Row7;
-import org.jooq.impl.UpdatableRecordImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class JGenerationLogsRecord extends UpdatableRecordImpl<JGenerationLogsRecord> implements Record7<UUID, UUID, Integer, String, String, String, LocalDateTime> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * Setter for <code>public.generation_logs.id</code>.
-     */
-    public void setId(UUID value) {
-        set(0, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.id</code>.
-     */
-    public UUID getId() {
-        return (UUID) get(0);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.report_id</code>.
-     */
-    public void setReportId(UUID value) {
-        set(1, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.report_id</code>.
-     */
-    public UUID getReportId() {
-        return (UUID) get(1);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.generation_number</code>.
-     */
-    public void setGenerationNumber(Integer value) {
-        set(2, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.generation_number</code>.
-     */
-    public Integer getGenerationNumber() {
-        return (Integer) get(2);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.prompt</code>.
-     */
-    public void setPrompt(String value) {
-        set(3, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.prompt</code>.
-     */
-    public String getPrompt() {
-        return (String) get(3);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.model_response</code>.
-     */
-    public void setModelResponse(String value) {
-        set(4, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.model_response</code>.
-     */
-    public String getModelResponse() {
-        return (String) get(4);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.user_feedback</code>.
-     */
-    public void setUserFeedback(String value) {
-        set(5, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.user_feedback</code>.
-     */
-    public String getUserFeedback() {
-        return (String) get(5);
-    }
-
-    /**
-     * Setter for <code>public.generation_logs.created_at</code>.
-     */
-    public void setCreatedAt(LocalDateTime value) {
-        set(6, value);
-    }
-
-    /**
-     * Getter for <code>public.generation_logs.created_at</code>.
-     */
-    public LocalDateTime getCreatedAt() {
-        return (LocalDateTime) get(6);
-    }
-
-    // -------------------------------------------------------------------------
-    // Primary key information
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Record1<UUID> key() {
-        return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record7 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row7<UUID, UUID, Integer, String, String, String, LocalDateTime> fieldsRow() {
-        return (Row7) super.fieldsRow();
-    }
-
-    @Override
-    public Row7<UUID, UUID, Integer, String, String, String, LocalDateTime> valuesRow() {
-        return (Row7) super.valuesRow();
-    }
-
-    @Override
-    public Field<UUID> field1() {
-        return JGenerationLogs.GENERATION_LOGS.ID;
-    }
-
-    @Override
-    public Field<UUID> field2() {
-        return JGenerationLogs.GENERATION_LOGS.REPORT_ID;
-    }
-
-    @Override
-    public Field<Integer> field3() {
-        return JGenerationLogs.GENERATION_LOGS.GENERATION_NUMBER;
-    }
-
-    @Override
-    public Field<String> field4() {
-        return JGenerationLogs.GENERATION_LOGS.PROMPT;
-    }
-
-    @Override
-    public Field<String> field5() {
-        return JGenerationLogs.GENERATION_LOGS.MODEL_RESPONSE;
-    }
-
-    @Override
-    public Field<String> field6() {
-        return JGenerationLogs.GENERATION_LOGS.USER_FEEDBACK;
-    }
-
-    @Override
-    public Field<LocalDateTime> field7() {
-        return JGenerationLogs.GENERATION_LOGS.CREATED_AT;
-    }
-
-    @Override
-    public UUID component1() {
-        return getId();
-    }
-
-    @Override
-    public UUID component2() {
-        return getReportId();
-    }
-
-    @Override
-    public Integer component3() {
-        return getGenerationNumber();
-    }
-
-    @Override
-    public String component4() {
-        return getPrompt();
-    }
-
-    @Override
-    public String component5() {
-        return getModelResponse();
-    }
-
-    @Override
-    public String component6() {
-        return getUserFeedback();
-    }
-
-    @Override
-    public LocalDateTime component7() {
-        return getCreatedAt();
-    }
-
-    @Override
-    public UUID value1() {
-        return getId();
-    }
-
-    @Override
-    public UUID value2() {
-        return getReportId();
-    }
-
-    @Override
-    public Integer value3() {
-        return getGenerationNumber();
-    }
-
-    @Override
-    public String value4() {
-        return getPrompt();
-    }
-
-    @Override
-    public String value5() {
-        return getModelResponse();
-    }
-
-    @Override
-    public String value6() {
-        return getUserFeedback();
-    }
-
-    @Override
-    public LocalDateTime value7() {
-        return getCreatedAt();
-    }
-
-    @Override
-    public JGenerationLogsRecord value1(UUID value) {
-        setId(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value2(UUID value) {
-        setReportId(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value3(Integer value) {
-        setGenerationNumber(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value4(String value) {
-        setPrompt(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value5(String value) {
-        setModelResponse(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value6(String value) {
-        setUserFeedback(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord value7(LocalDateTime value) {
-        setCreatedAt(value);
-        return this;
-    }
-
-    @Override
-    public JGenerationLogsRecord values(UUID value1, UUID value2, Integer value3, String value4, String value5, String value6, LocalDateTime value7) {
-        value1(value1);
-        value2(value2);
-        value3(value3);
-        value4(value4);
-        value5(value5);
-        value6(value6);
-        value7(value7);
-        return this;
-    }
-
-    // -------------------------------------------------------------------------
-    // Constructors
-    // -------------------------------------------------------------------------
-
-    /**
-     * Create a detached JGenerationLogsRecord
-     */
-    public JGenerationLogsRecord() {
-        super(JGenerationLogs.GENERATION_LOGS);
-    }
-
-    /**
-     * Create a detached, initialised JGenerationLogsRecord
-     */
-    public JGenerationLogsRecord(UUID id, UUID reportId, Integer generationNumber, String prompt, String modelResponse, String userFeedback, LocalDateTime createdAt) {
-        super(JGenerationLogs.GENERATION_LOGS);
-
-        setId(id);
-        setReportId(reportId);
-        setGenerationNumber(generationNumber);
-        setPrompt(prompt);
-        setModelResponse(modelResponse);
-        setUserFeedback(userFeedback);
-        setCreatedAt(createdAt);
-        resetChangedOnNotNull();
-    }
-}
