+++ backend/src/main/java/com/example/backend/application/usecases/reports/ReportUseCase.java
@@ -52,7 +52,6 @@ public class ReportUseCase {
                 .reportDate(existingReport.getReportDate())
                 .rawData(existingReport.getRawData())
                 .finalContent(request.getFinalContent())
-                .generationCount(existingReport.getGenerationCount())
                 .additionalNotes(request.getAdditionalNotes() != null ? request.getAdditionalNotes() : existingReport.getAdditionalNotes())
                 .createdAt(existingReport.getCreatedAt())
                 .updatedAt(LocalDateTime.now())
@@ -132,7 +131,6 @@ public class ReportUseCase {
                 .reportDate(report.getReportDate())
                 .rawData(report.getRawData())
                 .finalContent(report.getFinalContent())
-                .generationCount(report.getGenerationCount())
                 .additionalNotes(report.getAdditionalNotes())
                 .createdAt(report.getCreatedAt())
                 .updatedAt(report.getUpdatedAt())
