+++ backend/src/main/java/com/example/backend/infrastructure/repositories/reports/DailyReportRepository.java
@@ -131,7 +131,6 @@ public class DailyReportRepository implements IDailyReportRepository {
                 .set(DAILY_REPORTS.REPORT_DATE, report.getReportDate())
                 .set(DAILY_REPORTS.RAW_DATA, report.getRawData() != null ? JSONB.valueOf(report.getRawData()) : null)
                 .set(DAILY_REPORTS.FINAL_CONTENT, report.getFinalContent())
-                .set(DAILY_REPORTS.GENERATION_COUNT, report.getGenerationCount())
                 .set(DAILY_REPORTS.ADDITIONAL_NOTES, report.getAdditionalNotes())
                 .set(DAILY_REPORTS.CREATED_AT, report.getCreatedAt())
                 .set(DAILY_REPORTS.UPDATED_AT, report.getUpdatedAt())
@@ -151,7 +150,6 @@ public class DailyReportRepository implements IDailyReportRepository {
         dsl.update(DAILY_REPORTS)
                 .set(DAILY_REPORTS.RAW_DATA, report.getRawData() != null ? JSONB.valueOf(report.getRawData()) : null)
                 .set(DAILY_REPORTS.FINAL_CONTENT, report.getFinalContent())
-                .set(DAILY_REPORTS.GENERATION_COUNT, report.getGenerationCount())
                 .set(DAILY_REPORTS.ADDITIONAL_NOTES, report.getAdditionalNotes())
                 .set(DAILY_REPORTS.UPDATED_AT, LocalDateTime.now())
                 .where(DAILY_REPORTS.ID.eq(report.getId()))
@@ -175,7 +173,6 @@ public class DailyReportRepository implements IDailyReportRepository {
                 .reportDate(record.get(DAILY_REPORTS.REPORT_DATE))
                 .rawData(record.get(DAILY_REPORTS.RAW_DATA) != null ? record.get(DAILY_REPORTS.RAW_DATA).toString() : null)
                 .finalContent(record.get(DAILY_REPORTS.FINAL_CONTENT))
-                .generationCount(record.get(DAILY_REPORTS.GENERATION_COUNT))
                 .additionalNotes(record.get(DAILY_REPORTS.ADDITIONAL_NOTES))
                 .createdAt(record.get(DAILY_REPORTS.CREATED_AT))
                 .updatedAt(record.get(DAILY_REPORTS.UPDATED_AT))
