+++ docs/database-design.md
@@ -31,11 +31,7 @@
 | user_id | UUID | FOREIGN KEY | ユーザーID |
 | report_date | DATE | NOT NULL | 日報対象日 |
 | raw_data | JSONB | | 収集した生データ |
-| generated_content | TEXT | | AI生成された日報内容 |
-| edited_content | TEXT | | 編集後の日報内容 |
 | final_content | TEXT | | 最終的な日報内容 |
-| status | VARCHAR(20) | NOT NULL | ステータス(DRAFT/EDITED/APPROVED) |
-| generation_count | INTEGER | DEFAULT 0 | 生成回数 |
 | additional_notes | TEXT | | ユーザー追加情報 |
 | created_at | TIMESTAMP | NOT NULL | 作成日時 |
 | updated_at | TIMESTAMP | NOT NULL | 更新日時 |
@@ -51,16 +47,6 @@
 | created_at | TIMESTAMP | NOT NULL | 作成日時 |
 | updated_at | TIMESTAMP | NOT NULL | 更新日時 |
 
-### 5. generation_logs (生成ログ)
-| カラム名 | 型 | 制約 | 説明 |
-|---------|-----|------|------|
-| id | UUID | PRIMARY KEY | ログID |
-| report_id | UUID | FOREIGN KEY | 日報ID |
-| generation_number | INTEGER | NOT NULL | 生成回数 |
-| prompt | TEXT | | 使用したプロンプト |
-| model_response | TEXT | | AIのレスポンス |
-| user_feedback | TEXT | | ユーザーフィードバック |
-| created_at | TIMESTAMP | NOT NULL | 生成日時 |
 
 ## インデックス
 
@@ -81,8 +67,6 @@ CREATE UNIQUE INDEX idx_daily_reports_user_date ON daily_reports(user_id, report
 -- report_templates
 CREATE INDEX idx_report_templates_user_id ON report_templates(user_id);
 
--- generation_logs
-CREATE INDEX idx_generation_logs_report_id ON generation_logs(report_id);
 ```
 
 ## 制約
